fix(zod): enforce integer validation

+++ samples/swr-with-zod/__snapshots__/models/labradoodle-reusable.zod.ts
@@ -7,7 +7,7 @@
 import { z as zod } from 'zod';
 
 export const Labradoodle = zod.object({
-  cuteness: zod.number(),
+  cuteness: zod.number().int(),
   breed: zod.enum(['Labradoodle']),
 });
